feat: use orderinfolib and remove ordervalidator
We don't actually need to store an orderHash mapping to assert an order
isn't filled twice - permitpost handles this with its nonce replay
protection already. It was just a helper function but traded off a lot
of gas as it was an always-cold sstore. This commit removes the
OrderValidator entirely, in favor of an OrderInfoLib with static
validation functions