feat: catch invalid decay
If startAmount < endAmount, dutch resolution currently fails with
underflow because we subtract startAmount from endAmount to get the
decay. This is generally an OK assumption as general usage of dutch
orders starts with a good price (high output) and over time gets worse
(lower output). This commit just adds a nicer error to catch this