feat: add indexed fields to fill event
This commit adds indexed fields to the Fill event -- orderHash, filler
and nonce. This adds a bit of gas overhead as indexed fields are more
expensive than non-indexed. orderHash seems a bit strange to index on since
it's always unique so it acts more as a lookup than a filter. Needs
a bit more thought - will chew on this over the weekend