feat: remove permit2 dependency and remove via-ir
viaIR compilation is a drag on development as it slows down compilation
/ testing times by a ton. It is currently required because of the
permit2 dependency. This commit removes the permit2 dependency by
etching bytecode and using a local interface for permit2 instead.
Compilation time down to 7 seconds

TODO: use direct permit2 dependency in future if/when viaIR speedups come

Downside of this change is that we may become out-of-sync with permit2
more easily